Layout, Nav, and SidebarItems converted to tsx
