+++ src/components/Layout.tsx
@@ -3,13 +3,13 @@ import React from 'react';
 import Sidebar from "./Sidebar";
 import Nav from "./Nav";
 
-function Layout(props) {
+function Layout(props: any) { // fix any as we have expected fields below
     return (
         <div>
             <div style={{display: "flex"}}>
                 <Sidebar history={props.history}/>
                 <div style={{maxWidth: '800px'}}>
-                    <Nav/>
+                    {/* <Nav/> */}
                     {props.children}
                 </div>
             </div>
