+++ src/components/Nav.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Nav() {
+  return (
+    <div>
+      <h2>I'm a nav</h2>
+      <p>Ignore me. I'm not important. I'm cool, though.</p>
+    </div>
+  );
+}
+
+export default Nav;
