--- src/components/SidebarItems.js
@@ -1,24 +1,0 @@
-const SidebarItems = [
-    {
-        name: "THE ITALIAN JOB",
-        route: '/'
-    },
-    {
-        name: "Dashboard",
-        route: '/dashboard',
-    },
-    {
-        name: "Page 1",
-        route: '/page-1'
-    },
-    {
-        name: "Page 2",
-        route: '/page-2'
-    },
-    {
-        name: "Page 3",
-        route: 'page-3'
-    },
-];
-
-export default SidebarItems;
